Network.AWS.ApplicationAutoScaling.DeregisterScalableTarget

Description

Deregisters a scalable target.

Deregistering a scalable target deletes the scaling policies that are associated with it.

To create a scalable target or update an existing one, see RegisterScalableTarget .

Creating a Request

deregisterScalableTarget

Arguments

:: ServiceNamespace
derServiceNamespace

-> Text
derResourceId

-> ScalableDimension
derScalableDimension

-> DeregisterScalableTarget

Creates a value of DeregisterScalableTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

derServiceNamespace - The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

derResourceId - The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

derScalableDimension - The scalable dimension associated with the scalable target. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

data DeregisterScalableTarget

See: deregisterScalableTarget smart constructor.

Request Lenses

derServiceNamespace :: Lens' DeregisterScalableTarget ServiceNamespace

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

derResourceId :: Lens' DeregisterScalableTarget Text

The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

derScalableDimension :: Lens' DeregisterScalableTarget ScalableDimension

The scalable dimension associated with the scalable target. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

Destructuring the Response

deregisterScalableTargetResponse

Arguments

:: Int
dstrsResponseStatus

-> DeregisterScalableTargetResponse

Creates a value of DeregisterScalableTargetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dstrsResponseStatus - -- | The response status code.

data DeregisterScalableTargetResponse

See: deregisterScalableTargetResponse smart constructor.

Response Lenses

dstrsResponseStatus :: Lens' DeregisterScalableTargetResponse Int

- | The response status code.