Network.AWS.ApplicationAutoScaling.DescribeScalableTargets

Description

Provides descriptive information about the scalable targets in the specified namespace.

You can filter the results using the ResourceIds and ScalableDimension parameters.

To create a scalable target or update an existing one, see RegisterScalableTarget . If you are no longer using a scalable target, you can deregister it using DeregisterScalableTarget .

This operation returns paginated results.

Creating a Request

describeScalableTargets

Arguments

:: ServiceNamespace
dstServiceNamespace

-> DescribeScalableTargets

Creates a value of DescribeScalableTargets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dstResourceIds - The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier. If you specify a scalable dimension, you must also specify a resource ID. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

dstScalableDimension - The scalable dimension associated with the scalable target. This string consists of the service namespace, resource type, and scaling property. If you specify a scalable dimension, you must also specify a resource ID. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

dstNextToken - The token for the next set of results.

dstMaxResults - The maximum number of scalable target results. This value can be between 1 and 50. The default value is 50. If this parameter is used, the operation returns up to MaxResults results at a time, along with a NextToken value. To get the next set of results, include the NextToken value in a subsequent call. If this parameter is not used, the operation returns up to 50 results and a NextToken value, if applicable.

dstServiceNamespace - The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

data DescribeScalableTargets

See: describeScalableTargets smart constructor.

Request Lenses

dstResourceIds :: Lens' DescribeScalableTargets [Text]

The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier. If you specify a scalable dimension, you must also specify a resource ID. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

dstScalableDimension :: Lens' DescribeScalableTargets (Maybe ScalableDimension)

The scalable dimension associated with the scalable target. This string consists of the service namespace, resource type, and scaling property. If you specify a scalable dimension, you must also specify a resource ID. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

dstNextToken :: Lens' DescribeScalableTargets (Maybe Text)

The token for the next set of results.

dstMaxResults :: Lens' DescribeScalableTargets (Maybe Int)

The maximum number of scalable target results. This value can be between 1 and 50. The default value is 50. If this parameter is used, the operation returns up to MaxResults results at a time, along with a NextToken value. To get the next set of results, include the NextToken value in a subsequent call. If this parameter is not used, the operation returns up to 50 results and a NextToken value, if applicable.

dstServiceNamespace :: Lens' DescribeScalableTargets ServiceNamespace

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

Destructuring the Response

describeScalableTargetsResponse

Arguments

:: Int
dstsrsResponseStatus

-> DescribeScalableTargetsResponse

Creates a value of DescribeScalableTargetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dstsrsNextToken - The token required to get the next set of results. This value is null if there are no more results to return.

dstsrsScalableTargets - The list of scalable targets that matches the request parameters.

dstsrsResponseStatus - -- | The response status code.

data DescribeScalableTargetsResponse

See: describeScalableTargetsResponse smart constructor.

Response Lenses

dstsrsNextToken :: Lens' DescribeScalableTargetsResponse (Maybe Text)

The token required to get the next set of results. This value is null if there are no more results to return.

dstsrsScalableTargets :: Lens' DescribeScalableTargetsResponse [ScalableTarget]

The list of scalable targets that matches the request parameters.

dstsrsResponseStatus :: Lens' DescribeScalableTargetsResponse Int

- | The response status code.