Network.AWS.ApplicationAutoScaling.DescribeScalingPolicies

Description

Provides descriptive information about the scaling policies in the specified namespace.

You can filter the results using the ResourceId , ScalableDimension , and PolicyNames parameters.

To create a scaling policy or update an existing one, see PutScalingPolicy . If you are no longer using a scaling policy, you can delete it using DeleteScalingPolicy .

This operation returns paginated results.

Creating a Request

describeScalingPolicies

Arguments

:: ServiceNamespace
dServiceNamespace

-> DescribeScalingPolicies

Creates a value of DescribeScalingPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dPolicyNames - The names of the scaling policies to describe.

dScalableDimension - The scalable dimension. This string consists of the service namespace, resource type, and scaling property. If you specify a scalable dimension, you must also specify a resource ID. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

dResourceId - The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier. If you specify a scalable dimension, you must also specify a resource ID. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

dNextToken - The token for the next set of results.

dMaxResults - The maximum number of scalable target results. This value can be between 1 and 50. The default value is 50. If this parameter is used, the operation returns up to MaxResults results at a time, along with a NextToken value. To get the next set of results, include the NextToken value in a subsequent call. If this parameter is not used, the operation returns up to 50 results and a NextToken value, if applicable.

dServiceNamespace - The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

data DescribeScalingPolicies

See: describeScalingPolicies smart constructor.

Request Lenses

dPolicyNames :: Lens' DescribeScalingPolicies [Text]

The names of the scaling policies to describe.

dScalableDimension :: Lens' DescribeScalingPolicies (Maybe ScalableDimension)

The scalable dimension. This string consists of the service namespace, resource type, and scaling property. If you specify a scalable dimension, you must also specify a resource ID. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

dResourceId :: Lens' DescribeScalingPolicies (Maybe Text)

The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier. If you specify a scalable dimension, you must also specify a resource ID. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

dNextToken :: Lens' DescribeScalingPolicies (Maybe Text)

The token for the next set of results.

dMaxResults :: Lens' DescribeScalingPolicies (Maybe Int)

The maximum number of scalable target results. This value can be between 1 and 50. The default value is 50. If this parameter is used, the operation returns up to MaxResults results at a time, along with a NextToken value. To get the next set of results, include the NextToken value in a subsequent call. If this parameter is not used, the operation returns up to 50 results and a NextToken value, if applicable.

dServiceNamespace :: Lens' DescribeScalingPolicies ServiceNamespace

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

Destructuring the Response

describeScalingPoliciesResponse

Arguments

:: Int
drsResponseStatus

-> DescribeScalingPoliciesResponse

Creates a value of DescribeScalingPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsNextToken - The token required to get the next set of results. This value is null if there are no more results to return.

drsScalingPolicies - A list of scaling policy objects.

drsResponseStatus - -- | The response status code.

data DescribeScalingPoliciesResponse

See: describeScalingPoliciesResponse smart constructor.

Response Lenses

drsNextToken :: Lens' DescribeScalingPoliciesResponse (Maybe Text)

The token required to get the next set of results. This value is null if there are no more results to return.

drsScalingPolicies :: Lens' DescribeScalingPoliciesResponse [ScalingPolicy]

A list of scaling policy objects.

drsResponseStatus :: Lens' DescribeScalingPoliciesResponse Int

- | The response status code.