Network.AWS.ApplicationAutoScaling.PutScalingPolicy

Description

Creates or updates a policy for an Application Auto Scaling scalable target.

Each scalable target is identified by a service namespace, resource ID, and scalable dimension. A scaling policy applies to the scalable target identified by those three attributes. You cannot create a scaling policy without first registering a scalable target using RegisterScalableTarget .

To update a policy, specify its policy name and the parameters that you want to change. Any parameters that you don't specify are not changed by this update request.

You can view the scaling policies for a service namespace using DescribeScalingPolicies . If you are no longer using a scaling policy, you can delete it using DeleteScalingPolicy .

Creating a Request

putScalingPolicy

Arguments

:: Text
pspPolicyName

-> ServiceNamespace
pspServiceNamespace

-> Text
pspResourceId

-> ScalableDimension
pspScalableDimension

-> PutScalingPolicy

Creates a value of PutScalingPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pspPolicyType - The policy type. If you are creating a new policy, this parameter is required. If you are updating a policy, this parameter is not required. For DynamoDB, only TargetTrackingScaling is supported. For any other service, only StepScaling is supported.

pspTargetTrackingScalingPolicyConfiguration - A target tracking policy. This parameter is required if you are creating a new policy and the policy type is TargetTrackingScaling .

pspStepScalingPolicyConfiguration - A step scaling policy. This parameter is required if you are creating a policy and the policy type is StepScaling .

pspPolicyName - The name of the scaling policy.

pspServiceNamespace - The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

pspResourceId - The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

pspScalableDimension - The scalable dimension. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

data PutScalingPolicy

See: putScalingPolicy smart constructor.

Request Lenses

pspPolicyType :: Lens' PutScalingPolicy (Maybe PolicyType)

The policy type. If you are creating a new policy, this parameter is required. If you are updating a policy, this parameter is not required. For DynamoDB, only TargetTrackingScaling is supported. For any other service, only StepScaling is supported.

pspTargetTrackingScalingPolicyConfiguration :: Lens' PutScalingPolicy (Maybe TargetTrackingScalingPolicyConfiguration)

A target tracking policy. This parameter is required if you are creating a new policy and the policy type is TargetTrackingScaling .

pspStepScalingPolicyConfiguration :: Lens' PutScalingPolicy (Maybe StepScalingPolicyConfiguration)

A step scaling policy. This parameter is required if you are creating a policy and the policy type is StepScaling .

pspPolicyName :: Lens' PutScalingPolicy Text

The name of the scaling policy.

pspServiceNamespace :: Lens' PutScalingPolicy ServiceNamespace

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

pspResourceId :: Lens' PutScalingPolicy Text

The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

pspScalableDimension :: Lens' PutScalingPolicy ScalableDimension

The scalable dimension. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

Destructuring the Response

putScalingPolicyResponse

Arguments

:: Int
psprsResponseStatus

-> Text
psprsPolicyARN

-> PutScalingPolicyResponse

Creates a value of PutScalingPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psprsAlarms - The CloudWatch alarms created for the target tracking policy.

psprsResponseStatus - -- | The response status code.

psprsPolicyARN - The Amazon Resource Name (ARN) of the resulting scaling policy.

data PutScalingPolicyResponse

See: putScalingPolicyResponse smart constructor.

Response Lenses

psprsAlarms :: Lens' PutScalingPolicyResponse [Alarm]

The CloudWatch alarms created for the target tracking policy.

psprsResponseStatus :: Lens' PutScalingPolicyResponse Int

- | The response status code.

psprsPolicyARN :: Lens' PutScalingPolicyResponse Text

The Amazon Resource Name (ARN) of the resulting scaling policy.