Network.AWS.ApplicationAutoScaling.Types

Description

Service Configuration

applicationAutoScaling :: Service

API version 2016-02-06 of the Amazon Application Auto Scaling SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

An exception was thrown for a validation issue. Review the available parameters for the API request.

_FailedResourceAccessException :: AsError a => Getting (First ServiceError) a ServiceError

Failed access to resources caused an exception. This exception is thrown when Application Auto Scaling is unable to retrieve the alarms associated with a scaling policy due to a client error, for example, if the role ARN specified for a scalable target does not have permission to call the CloudWatch DescribeAlarms API operation on behalf of your account.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The next token supplied was invalid.

_ConcurrentUpdateException :: AsError a => Getting (First ServiceError) a ServiceError

Concurrent updates caused an exception, for example, if you request an update to an Application Auto Scaling resource that already has a pending update.

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The service encountered an internal error.

_ObjectNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified object could not be found. For any Put or Register API operation, which depends on the existence of a scalable target, this exception is thrown if the scalable target with the specified service namespace, resource ID, and scalable dimension does not exist. For any Delete or Deregister API operation, this exception is thrown if the resource that is to be deleted or deregistered cannot be found.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your account exceeded a limit. This exception is thrown when a per-account resource limit is exceeded. For more information, see Application Auto Scaling Limits .

AdjustmentType

data AdjustmentType

Constructors

ChangeInCapacity

ExactCapacity

PercentChangeInCapacity

MetricAggregationType

data MetricAggregationType

Constructors

MATAverage

MATMaximum

MATMinimum

MetricStatistic

data MetricStatistic

Constructors

Average

Maximum

Minimum

SampleCount

Sum

MetricType

data MetricType

Constructors

DynamoDBReadCapacityUtilization

DynamoDBWriteCapacityUtilization

PolicyType

data PolicyType

Constructors

StepScaling

TargetTrackingScaling

ScalableDimension

data ScalableDimension

Constructors

AppstreamFleetDesiredCapacity

DynamodbIndexReadCapacityUnits

DynamodbIndexWriteCapacityUnits

DynamodbTableReadCapacityUnits

DynamodbTableWriteCapacityUnits

EC2SpotFleetRequestTargetCapacity

EcsServiceDesiredCount

ElasticmapreduceInstancegroupInstanceCount

ScalingActivityStatusCode

data ScalingActivityStatusCode

Constructors

Failed

InProgress

Overridden

Pending

Successful

Unfulfilled

ServiceNamespace

data ServiceNamespace

Constructors

Appstream

Dynamodb

EC2

Ecs

Elasticmapreduce

Alarm

data Alarm

Represents a CloudWatch alarm associated with a scaling policy.

See: alarm smart constructor.

alarm

Arguments

:: Text
aAlarmName

-> Text
aAlarmARN

-> Alarm

Creates a value of Alarm with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aAlarmName - The name of the alarm.

aAlarmARN - The Amazon Resource Name (ARN) of the alarm.

aAlarmName :: Lens' Alarm Text

The name of the alarm.

aAlarmARN :: Lens' Alarm Text

The Amazon Resource Name (ARN) of the alarm.

CustomizedMetricSpecification

data CustomizedMetricSpecification

Configures a customized metric for a target tracking policy.

See: customizedMetricSpecification smart constructor.

customizedMetricSpecification

Arguments

:: Text
cmsMetricName

-> Text
cmsNamespace

-> MetricStatistic
cmsStatistic

-> CustomizedMetricSpecification

Creates a value of CustomizedMetricSpecification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmsDimensions - The dimensions of the metric.

cmsUnit - The unit of the metric.

cmsMetricName - The name of the metric.

cmsNamespace - The namespace of the metric.

cmsStatistic - The statistic of the metric.

cmsDimensions :: Lens' CustomizedMetricSpecification [MetricDimension]

The dimensions of the metric.

cmsUnit :: Lens' CustomizedMetricSpecification (Maybe Text)

The unit of the metric.

cmsMetricName :: Lens' CustomizedMetricSpecification Text

The name of the metric.

cmsNamespace :: Lens' CustomizedMetricSpecification Text

The namespace of the metric.

cmsStatistic :: Lens' CustomizedMetricSpecification MetricStatistic

The statistic of the metric.

MetricDimension

data MetricDimension

Describes the dimension of a metric.

See: metricDimension smart constructor.

metricDimension

Arguments

:: Text
mdName

-> Text
mdValue

-> MetricDimension

Creates a value of MetricDimension with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdName - The name of the dimension.

mdValue - The value of the dimension.

mdName :: Lens' MetricDimension Text

The name of the dimension.

mdValue :: Lens' MetricDimension Text

The value of the dimension.

PredefinedMetricSpecification

data PredefinedMetricSpecification

Configures a predefined metric for a target tracking policy.

See: predefinedMetricSpecification smart constructor.

predefinedMetricSpecification

Arguments

:: MetricType
pmsPredefinedMetricType

-> PredefinedMetricSpecification

Creates a value of PredefinedMetricSpecification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pmsResourceLabel - Reserved for future use.

pmsPredefinedMetricType - The metric type.

pmsResourceLabel :: Lens' PredefinedMetricSpecification (Maybe Text)

Reserved for future use.

pmsPredefinedMetricType :: Lens' PredefinedMetricSpecification MetricType

The metric type.

ScalableTarget

data ScalableTarget

Represents a scalable target.

See: scalableTarget smart constructor.

scalableTarget

Arguments

:: ServiceNamespace
stServiceNamespace

-> Text
stResourceId

-> ScalableDimension
stScalableDimension

-> Int
stMinCapacity

-> Int
stMaxCapacity

-> Text
stRoleARN

-> UTCTime
stCreationTime

-> ScalableTarget

Creates a value of ScalableTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stServiceNamespace - The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

stResourceId - The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

stScalableDimension - The scalable dimension associated with the scalable target. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

stMinCapacity - The minimum value to scale to in response to a scale in event.

stMaxCapacity - The maximum value to scale to in response to a scale out event.

stRoleARN - The ARN of an IAM role that allows Application Auto Scaling to modify the scalable target on your behalf.

stCreationTime - The Unix timestamp for when the scalable target was created.

stServiceNamespace :: Lens' ScalableTarget ServiceNamespace

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

stResourceId :: Lens' ScalableTarget Text

The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

stScalableDimension :: Lens' ScalableTarget ScalableDimension

The scalable dimension associated with the scalable target. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

stMinCapacity :: Lens' ScalableTarget Int

The minimum value to scale to in response to a scale in event.

stMaxCapacity :: Lens' ScalableTarget Int

The maximum value to scale to in response to a scale out event.

stRoleARN :: Lens' ScalableTarget Text

The ARN of an IAM role that allows Application Auto Scaling to modify the scalable target on your behalf.

stCreationTime :: Lens' ScalableTarget UTCTime

The Unix timestamp for when the scalable target was created.

ScalingActivity

data ScalingActivity

Represents a scaling activity.

See: scalingActivity smart constructor.

scalingActivity

Arguments

:: Text
saActivityId

-> ServiceNamespace
saServiceNamespace

-> Text
saResourceId

-> ScalableDimension
saScalableDimension

-> Text
saDescription

-> Text
saCause

-> UTCTime
saStartTime

-> ScalingActivityStatusCode
saStatusCode

-> ScalingActivity

Creates a value of ScalingActivity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

saStatusMessage - A simple message about the current status of the scaling activity.

saEndTime - The Unix timestamp for when the scaling activity ended.

saDetails - The details about the scaling activity.

saActivityId - The unique identifier of the scaling activity.

saServiceNamespace - The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

saResourceId - The identifier of the resource associated with the scaling activity. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

saScalableDimension - The scalable dimension. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

saDescription - A simple description of what action the scaling activity intends to accomplish.

saCause - A simple description of what caused the scaling activity to happen.

saStartTime - The Unix timestamp for when the scaling activity began.

saStatusCode - Indicates the status of the scaling activity.

saStatusMessage :: Lens' ScalingActivity (Maybe Text)

A simple message about the current status of the scaling activity.

saEndTime :: Lens' ScalingActivity (Maybe UTCTime)

The Unix timestamp for when the scaling activity ended.

saDetails :: Lens' ScalingActivity (Maybe Text)

The details about the scaling activity.

saActivityId :: Lens' ScalingActivity Text

The unique identifier of the scaling activity.

saServiceNamespace :: Lens' ScalingActivity ServiceNamespace

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

saResourceId :: Lens' ScalingActivity Text

The identifier of the resource associated with the scaling activity. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

saScalableDimension :: Lens' ScalingActivity ScalableDimension

The scalable dimension. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

saDescription :: Lens' ScalingActivity Text

A simple description of what action the scaling activity intends to accomplish.

saCause :: Lens' ScalingActivity Text

A simple description of what caused the scaling activity to happen.

saStartTime :: Lens' ScalingActivity UTCTime

The Unix timestamp for when the scaling activity began.

saStatusCode :: Lens' ScalingActivity ScalingActivityStatusCode

Indicates the status of the scaling activity.

ScalingPolicy

data ScalingPolicy

Represents a scaling policy.

See: scalingPolicy smart constructor.

scalingPolicy

Arguments

:: Text
spPolicyARN

-> Text
spPolicyName

-> ServiceNamespace
spServiceNamespace

-> Text
spResourceId

-> ScalableDimension
spScalableDimension

-> PolicyType
spPolicyType

-> UTCTime
spCreationTime

-> ScalingPolicy

Creates a value of ScalingPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spTargetTrackingScalingPolicyConfiguration - A target tracking policy.

spStepScalingPolicyConfiguration - A step scaling policy.

spAlarms - The CloudWatch alarms associated with the scaling policy.

spPolicyARN - The Amazon Resource Name (ARN) of the scaling policy.

spPolicyName - The name of the scaling policy.

spServiceNamespace - The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

spResourceId - The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

spScalableDimension - The scalable dimension. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

spPolicyType - The scaling policy type.

spCreationTime - The Unix timestamp for when the scaling policy was created.

spTargetTrackingScalingPolicyConfiguration :: Lens' ScalingPolicy (Maybe TargetTrackingScalingPolicyConfiguration)

A target tracking policy.

spStepScalingPolicyConfiguration :: Lens' ScalingPolicy (Maybe StepScalingPolicyConfiguration)

A step scaling policy.

spAlarms :: Lens' ScalingPolicy [Alarm]

The CloudWatch alarms associated with the scaling policy.

spPolicyARN :: Lens' ScalingPolicy Text

The Amazon Resource Name (ARN) of the scaling policy.

spPolicyName :: Lens' ScalingPolicy Text

The name of the scaling policy.

spServiceNamespace :: Lens' ScalingPolicy ServiceNamespace

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference .

spResourceId :: Lens' ScalingPolicy Text

The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier. * ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: servicedefaultsample-webapp . * Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE . * EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroupj-2EEZNYKUA1NTVig-1791Y4E1L8YI0 . * AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet . * DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table . * DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: tablemy-tableindex/my-table-index .

spScalableDimension :: Lens' ScalingPolicy ScalableDimension

The scalable dimension. This string consists of the service namespace, resource type, and scaling property. * ecs:service:DesiredCount - The desired task count of an ECS service. * ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request. * elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group. * appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet. * dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table. * dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table. * dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index. * dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

spPolicyType :: Lens' ScalingPolicy PolicyType

The scaling policy type.

spCreationTime :: Lens' ScalingPolicy UTCTime

The Unix timestamp for when the scaling policy was created.

StepAdjustment

data StepAdjustment

Represents a step adjustment for a StepScalingPolicyConfiguration . Describes an adjustment based on the difference between the value of the aggregated CloudWatch metric and the breach threshold that you've defined for the alarm.

For the following examples, suppose that you have an alarm with a breach threshold of 50:

To trigger the adjustment when the metric is greater than or equal to 50 and less than 60, specify a lower bound of 0 and an upper bound of 10.

To trigger the adjustment when the metric is greater than 40 and less than or equal to 50, specify a lower bound of -10 and an upper bound of 0.

There are a few rules for the step adjustments for your step policy:

The ranges of your step adjustments can't overlap or have a gap.

At most one step adjustment can have a null lower bound. If one step adjustment has a negative lower bound, then there must be a step adjustment with a null lower bound.

At most one step adjustment can have a null upper bound. If one step adjustment has a positive upper bound, then there must be a step adjustment with a null upper bound.

The upper and lower bound can't be null in the same step adjustment.

See: stepAdjustment smart constructor.

stepAdjustment

Arguments

:: Int
saScalingAdjustment

-> StepAdjustment

Creates a value of StepAdjustment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

saMetricIntervalLowerBound - The lower bound for the difference between the alarm threshold and the CloudWatch metric. If the metric value is above the breach threshold, the lower bound is inclusive (the metric must be greater than or equal to the threshold plus the lower bound). Otherwise, it is exclusive (the metric must be greater than the threshold plus the lower bound). A null value indicates negative infinity.

saMetricIntervalUpperBound - The upper bound for the difference between the alarm threshold and the CloudWatch metric. If the metric value is above the breach threshold, the upper bound is exclusive (the metric must be less than the threshold plus the upper bound). Otherwise, it is inclusive (the metric must be less than or equal to the threshold plus the upper bound). A null value indicates positive infinity. The upper bound must be greater than the lower bound.

saScalingAdjustment - The amount by which to scale, based on the specified adjustment type. A positive value adds to the current scalable dimension while a negative number removes from the current scalable dimension.

saMetricIntervalLowerBound :: Lens' StepAdjustment (Maybe Double)

The lower bound for the difference between the alarm threshold and the CloudWatch metric. If the metric value is above the breach threshold, the lower bound is inclusive (the metric must be greater than or equal to the threshold plus the lower bound). Otherwise, it is exclusive (the metric must be greater than the threshold plus the lower bound). A null value indicates negative infinity.

saMetricIntervalUpperBound :: Lens' StepAdjustment (Maybe Double)

The upper bound for the difference between the alarm threshold and the CloudWatch metric. If the metric value is above the breach threshold, the upper bound is exclusive (the metric must be less than the threshold plus the upper bound). Otherwise, it is inclusive (the metric must be less than or equal to the threshold plus the upper bound). A null value indicates positive infinity. The upper bound must be greater than the lower bound.

saScalingAdjustment :: Lens' StepAdjustment Int

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current scalable dimension while a negative number removes from the current scalable dimension.

StepScalingPolicyConfiguration

data StepScalingPolicyConfiguration

Represents a step scaling policy configuration.

See: stepScalingPolicyConfiguration smart constructor.

stepScalingPolicyConfiguration :: StepScalingPolicyConfiguration

Creates a value of StepScalingPolicyConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sspcStepAdjustments - A set of adjustments that enable you to scale based on the size of the alarm breach.

sspcAdjustmentType - The adjustment type, which specifies how the ScalingAdjustment parameter in a StepAdjustment is interpreted.

sspcCooldown - The amount of time, in seconds, after a scaling activity completes where previous trigger-related scaling activities can influence future scaling events. For scale out policies, while the cooldown period is in effect, the capacity that has been added by the previous scale out event that initiated the cooldown is calculated as part of the desired capacity for the next scale out. The intention is to continuously (but not excessively) scale out. For example, an alarm triggers a step scaling policy to scale out an Amazon ECS service by 2 tasks, the scaling activity completes successfully, and a cooldown period of 5 minutes starts. During the Cooldown period, if the alarm triggers the same policy again but at a more aggressive step adjustment to scale out the service by 3 tasks, the 2 tasks that were added in the previous scale out event are considered part of that capacity and only 1 additional task is added to the desired count. For scale in policies, the cooldown period is used to block subsequent scale in requests until it has expired. The intention is to scale in conservatively to protect your application's availability. However, if another alarm triggers a scale out policy during the cooldown period after a scale-in, Application Auto Scaling scales out your scalable target immediately.

sspcMetricAggregationType - The aggregation type for the CloudWatch metrics. Valid values are Minimum , Maximum , and Average .

sspcMinAdjustmentMagnitude - The minimum number to adjust your scalable dimension as a result of a scaling activity. If the adjustment type is PercentChangeInCapacity , the scaling policy changes the scalable dimension of the scalable target by this amount.

sspcStepAdjustments :: Lens' StepScalingPolicyConfiguration [StepAdjustment]

A set of adjustments that enable you to scale based on the size of the alarm breach.

sspcAdjustmentType :: Lens' StepScalingPolicyConfiguration (Maybe AdjustmentType)

The adjustment type, which specifies how the ScalingAdjustment parameter in a StepAdjustment is interpreted.

sspcCooldown :: Lens' StepScalingPolicyConfiguration (Maybe Int)

The amount of time, in seconds, after a scaling activity completes where previous trigger-related scaling activities can influence future scaling events. For scale out policies, while the cooldown period is in effect, the capacity that has been added by the previous scale out event that initiated the cooldown is calculated as part of the desired capacity for the next scale out. The intention is to continuously (but not excessively) scale out. For example, an alarm triggers a step scaling policy to scale out an Amazon ECS service by 2 tasks, the scaling activity completes successfully, and a cooldown period of 5 minutes starts. During the Cooldown period, if the alarm triggers the same policy again but at a more aggressive step adjustment to scale out the service by 3 tasks, the 2 tasks that were added in the previous scale out event are considered part of that capacity and only 1 additional task is added to the desired count. For scale in policies, the cooldown period is used to block subsequent scale in requests until it has expired. The intention is to scale in conservatively to protect your application's availability. However, if another alarm triggers a scale out policy during the cooldown period after a scale-in, Application Auto Scaling scales out your scalable target immediately.

sspcMetricAggregationType :: Lens' StepScalingPolicyConfiguration (Maybe MetricAggregationType)

The aggregation type for the CloudWatch metrics. Valid values are Minimum , Maximum , and Average .

sspcMinAdjustmentMagnitude :: Lens' StepScalingPolicyConfiguration (Maybe Int)

The minimum number to adjust your scalable dimension as a result of a scaling activity. If the adjustment type is PercentChangeInCapacity , the scaling policy changes the scalable dimension of the scalable target by this amount.

TargetTrackingScalingPolicyConfiguration

data TargetTrackingScalingPolicyConfiguration

Represents a target tracking scaling policy configuration.

See: targetTrackingScalingPolicyConfiguration smart constructor.

targetTrackingScalingPolicyConfiguration

Arguments

:: Double
ttspcTargetValue

-> TargetTrackingScalingPolicyConfiguration

Creates a value of TargetTrackingScalingPolicyConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ttspcPredefinedMetricSpecification - A predefined metric.

ttspcScaleInCooldown - The amount of time, in seconds, after a scale in activity completes before another scale in activity can start. The cooldown period is used to block subsequent scale in requests until it has expired. The intention is to scale in conservatively to protect your application's availability. However, if another alarm triggers a scale out policy during the cooldown period after a scale-in, Application Auto Scaling scales out your scalable target immediately.

ttspcCustomizedMetricSpecification - Reserved for future use.

ttspcScaleOutCooldown - The amount of time, in seconds, after a scale out activity completes before another scale out activity can start. While the cooldown period is in effect, the capacity that has been added by the previous scale out event that initiated the cooldown is calculated as part of the desired capacity for the next scale out. The intention is to continuously (but not excessively) scale out.

ttspcTargetValue - The target value for the metric. The range is 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2).

ttspcPredefinedMetricSpecification :: Lens' TargetTrackingScalingPolicyConfiguration (Maybe PredefinedMetricSpecification)

A predefined metric.

ttspcScaleInCooldown :: Lens' TargetTrackingScalingPolicyConfiguration (Maybe Int)

The amount of time, in seconds, after a scale in activity completes before another scale in activity can start. The cooldown period is used to block subsequent scale in requests until it has expired. The intention is to scale in conservatively to protect your application's availability. However, if another alarm triggers a scale out policy during the cooldown period after a scale-in, Application Auto Scaling scales out your scalable target immediately.

ttspcCustomizedMetricSpecification :: Lens' TargetTrackingScalingPolicyConfiguration (Maybe CustomizedMetricSpecification)

Reserved for future use.

ttspcScaleOutCooldown :: Lens' TargetTrackingScalingPolicyConfiguration (Maybe Int)

The amount of time, in seconds, after a scale out activity completes before another scale out activity can start. While the cooldown period is in effect, the capacity that has been added by the previous scale out event that initiated the cooldown is calculated as part of the desired capacity for the next scale out. The intention is to continuously (but not excessively) scale out.

ttspcTargetValue :: Lens' TargetTrackingScalingPolicyConfiguration Double

The target value for the metric. The range is 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2).